Dev

See merge request stepavi2/sem_prace_omo_smart_farm!19

+++ src/main/java/cz/cvut/fel/omo/smartfarm/model/products/Egg.java
@@ -1,8 +1,8 @@
 package cz.cvut.fel.omo.smartfarm.model.products;
 
 public class Egg extends Product {
-    public Egg(String name, int price, int weight) {
-        super(name, price, weight, ProductType.EGG);  // Передаем тип продукта
+    public Egg(String name, double price, int weight) {
+        super(name, price, weight, ProductType.EGG);
     }
 }
