+++ src/main/java/cz/cvut/fel/omo/smartfarm/model/products/Meat.java
@@ -1,7 +1,7 @@
 package cz.cvut.fel.omo.smartfarm.model.products;
 
 public class Meat extends Product {
-    public Meat(String name, int price, int weight) {
-        super(name, price, weight, ProductType.MEAT);  // Передаем тип продукта
+    public Meat(String name, double price, int weight) {
+        super(name, price, weight, ProductType.MEAT);
     }
 }
