+++ src/main/java/cz/cvut/fel/omo/smartfarm/model/products/Milk.java
@@ -1,7 +1,7 @@
 package cz.cvut.fel.omo.smartfarm.model.products;
 
 public class Milk extends Product {
-    public Milk(String name, int price, int weight) {
+    public Milk(String name, double price, int weight) {
         super(name, price, weight, ProductType.MILK);
     }
 }
